+++ CHANGELOG.md
@@ -38,11 +38,21 @@ The Koto project adheres to
   x = 1_000_000
   y = 0xff_aa_bb
   ```
+- Interpolated values can now be formatted with alternative representations.
+  - E.g.
+  ```koto
+  print '{15:b}'
+  # -> 1111
+  ```
+  - The `@debug` metakey has been added to allow for additional debug information
+    to be provided when formatting an object as a string.
 
 #### API
 
 - `Compiler::compile_ast` has been added, useful for tools that want to work with the AST
   after checking that it compiles correctly.
+- `DisplayContext::debug_enabled` has been added to allow native objects to provide
+  additional debug information when `KotoObject::display` is called.
 
 ### Changed